Update 3-11 (three-eleven).cpp

+++ 3-11 (three-eleven).cpp
@@ -1,7 +1,7 @@
 #include<stdio.h>
 int main(){
-	long long int N;
-	scanf("%lld",&N);
-	printf("%d %d",N%3,N%11);
+	int N;
+	scanf("%d",&N);
+	printf("%u %u",N%3,N%11);
 	return 0;
 }
